% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.5_game_over_functions.R
\name{game_over_cards}
\alias{game_over_cards}
\title{game over function}
\usage{
game_over_cards(input, output, cards)
}
\description{
game over function
}
\keyword{game_over_cards}
\keyword{internal}
