% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0_jaipur-package.R
\docType{package}
\name{jaipur-package}
\alias{jaipur-package}
\title{Use R for Playing the card Game Jaipur}
\description{
This is the R implementation of the card game Jaipur.
It is a two player game, but played on a single monitor.
}
\examples{
## jaipur:.start_jaipur("Player one", "Player two")
}
\author{
Anton Thielmann, Brandon Schwab
}
\keyword{package}
