% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{jaipur}
\alias{jaipur}
\alias{jaipur-package}
\title{jaipur: A package of the card game jaipiur}
\description{
The controls should be familiar to you: place the Coursor on one of the actionboxes and
use the left mouse button to perform the desired action.
}
\details{
Usage: shinyApp(ui = ui, server = server)
}
