% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.6_overallServer.R
\name{jaipur_server}
\alias{jaipur_server}
\title{Jaipur server function}
\usage{
jaipur_server(input, output, session)
}
\description{
Jaipur server function
}
\keyword{internal}
\keyword{jaipur_server}
