% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.4_overallUI.R
\docType{data}
\name{jaipur_ui}
\alias{jaipur_ui}
\title{jaipur UI function}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
jaipur_ui
}
\description{
jaipur UI function
}
\keyword{internal}
\keyword{jaipur_ui}
