% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1_output_server.R
\name{output_server}
\alias{output_server}
\title{Output server function}
\usage{
output_server(input, output, server, cards)
}
\description{
Output server function
}
\keyword{internal}
\keyword{output_server}
