% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.3_players_tab_server.R
\name{player1_tab_server}
\alias{player1_tab_server}
\title{Player1 tab server function}
\usage{
player1_tab_server(input, output, session, cards)
}
\description{
Player1 tab server function
}
\keyword{internal}
\keyword{player1_tab_server}
