% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1_colorfunc_and_rules.R
\name{playing_cards_plot}
\alias{playing_cards_plot}
\title{Plot your playing cards}
\usage{
playing_cards_plot(title, cards, plot_position, x_limit, cards_color)
}
\description{
Plot playing cards as textboxplots
}
\note{
insert plot position as list and color as a vector
}
