% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.4_set_up_cards_server.R
\name{set_up_cards_server}
\alias{set_up_cards_server}
\title{Set up cards server function}
\usage{
set_up_cards_server(input, output, session, cards)
}
\description{
Set up cards server function
}
\keyword{internal}
\keyword{set_up_cards_server}
