% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1_start_jaipur.R
\name{start_jaipur}
\alias{start_jaipur}
\title{Start the Game Jaipur}
\usage{
start_jaipur(x)
}
\arguments{
\item{x}{TRUE or FALSE. If TRUE the app will start in your default Browser.
If FALSE the app will start in the R Browser}
}
\description{
Jaipur is a two player card game.
}
\note{
For more FUN while playing call shinyalert::useShinyalert()
in your console
}
