% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.5_taking_server.R
\name{taking_server}
\alias{taking_server}
\title{taking server function}
\usage{
taking_server(input, output, session, cards, parent_session)
}
\description{
taking server function
}
\keyword{internal}
\keyword{taking_server}
