% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1_colorfunc_and_rules.R
\name{playing_cards_plot}
\alias{playing_cards_plot}
\title{Plot your playing cards}
\usage{
playing_cards_plot(title = "Plot Title", cards, x, y = 0.85,
  x_limit = c(0, 10), cards_color = "white")
}
\arguments{
\item{title}{= The Title of your plot. Default = "Plot Title".}

\item{cards}{= Playing Cards that will be plotted as character.}

\item{x}{list of vectors with min.and max. x-position of cards.}

\item{y}{y position: location of the top of the card.}

\item{x_limit}{set x scale limit. Default is c(0, 10).}

\item{cards_color}{colors of playing cards, default is white.}
}
\description{
Plot playing cards as textboxplots
}
\examples{
# Generate some playing cards
 my_playing_cards <- c("Ace", "King", "Seven")

 playing_cards_plot(title = "My Playing Cards",
 cards = my_playing_cards,
 x = list(c(0, 1.5),
 c(2, 3.5),
 c(4, 5.5)),
 cards_color = c("green", "blue", "red"))
}
