% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1_start_jaipur.R
\name{start_jaipur}
\alias{start_jaipur}
\title{Start the Game Jaipur}
\usage{
start_jaipur(Player_one = "Player 1", Player_two = "Player 2")
}
\arguments{
\item{Player_one}{Title of Player one's tabPanel}

\item{Player_two}{Title of Player two's tabPanel}
}
\description{
Jaipur is a two player card game.
}
\note{
For more FUN while playing play the game in full screen mode!
}
\examples{
#start_jaipur("Donald", "Barack")


}
